Add error messages

+++ FriendlyFisherman/BoundedContexts/Users/Users.Services/Implementation/PersonalMessagesService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using FriendlyFisherman.SharedKernel.Messages;
 using Users.Domain.Entities;
 using Users.Domain.EntityViewModels.PersonalMessage;
 using Users.Domain.Repositories;
@@ -31,6 +32,9 @@ namespace Users.Services.Implementation
             var response = new GetAllMessagesResponse();
             try
             {
+                if (String.IsNullOrWhiteSpace(request.SenderId))
+                    throw new Exception(ErrorMessages.InvalidId);
+
                 var result = _repo.GetAllMessagesBySenderId(request.SenderId);
                 response.Messages = result.Select(m => new PersonalMessageViewModel(m));
             }
@@ -53,6 +57,9 @@ namespace Users.Services.Implementation
             var response = new GetAllMessagesResponse();
             try
             {
+                if (String.IsNullOrWhiteSpace(request.ReceiverId))
+                    throw new Exception(ErrorMessages.InvalidId);
+
                 var result = _repo.GetAllMessagesByReceiverId(request.ReceiverId);
                 response.Messages = result.Select(m => new PersonalMessageViewModel(m));
             }
@@ -75,6 +82,9 @@ namespace Users.Services.Implementation
             var response = new GetMessageResponse();
             try
             {
+                if (String.IsNullOrWhiteSpace(request.MessageId))
+                    throw new Exception(ErrorMessages.InvalidId);
+
                 var result = _repo.GetMessageById(request.MessageId);
                 response.Message = new PersonalMessageViewModel(result);
             }
@@ -131,6 +141,9 @@ namespace Users.Services.Implementation
 
             try
             {
+                if (String.IsNullOrWhiteSpace(request.Message.Id))
+                    throw new Exception(ErrorMessages.InvalidId);
+
                 var message = _repo.GetMessageById(request.Message.Id);
                 message.Seen = request.Message.Seen;
 
@@ -156,6 +169,9 @@ namespace Users.Services.Implementation
 
             try
             {
+                if (String.IsNullOrWhiteSpace(request.MessageId))
+                    throw new Exception(ErrorMessages.InvalidId);
+
                 _repo.DeleteMessage(request.MessageId);
             }
             catch (Exception e)
